backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Telegram
python-telegram-bot==21.0.1
aiogram==3.4.1

# Validation and serialization
pydantic==2.6.1
pydantic-settings==2.2.0
email-validator==2.1.0.post1

# HTTP client
httpx==0.27.0
aiohttp==3.9.3

# Task queue (optional for V2)
celery==5.3.6

# Utilities
python-dotenv==1.0.1
pytz==2024.1

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.27.0

# Dev tools
black==24.2.0
isort==5.13.2
flake8==7.0.0
